requirements.txt
# Browser Agent Requirements
# =========================

# Core Dependencies
playwright>=1.40.0
aiohttp>=3.9.0
pyyaml>=6.0

# Data Handling
pydantic>=2.0.0
python-dotenv>=1.0.0

# Async Support
tenacity>=8.2.0

# Image Processing
pillow>=10.0.0

# API Server (optional)
fastapi>=0.109.0
uvicorn>=0.27.0

# Logging
structlog>=24.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
